Splitting toss3 host-configs files
The idea is to separate project specific and shareable configuration.

As for now, shareable information mainly consist in toolchain
configuration, such as compiler paths, and is placed in the .gitlab/conf
directory. More generaly, the aim is to have .gitlab containing
non-project specific gitlab configuration so that it can be easily moved
elsewhere later on (submodule ?).

Project specific configuration, such as compiling flags and target
release identification, remains at the same place.

shareable and project configuration are link by the directory layout
which remains the same : system_type/toolchain