Creation of a GitLab CI pipeline
The .gitlab-ci.yml file is used by GitLab CI as the main
configuration file. Using inclusion of files (include:) and
job properties extensions (extends:) the pipeline configuration
is split in severale places in a modular manner.

.gitlab-ci.yml containins project specific configuration
of the pipeline.

scripts/gitlab contains projects specific scripts for gitlab.

.gitlab/ci contains shareable configuration, such as
configuration of jobs for different clusters.